fiet-242 error handling kleding bestellen clublid

+++ app/Http/Livewire/Member/MyOrder.php
@@ -93,7 +93,12 @@ class MyOrder extends Component
     {
         $order = auth()->user()->orders()->where('id', $orderId)->first();
 
-        $order->delete();
+        if (!$order) {
+            return;
+        } else {
+            $order->delete();
+        }
+
 
         $this->orders = $this->getOrdersForUser();
     }
